- moving ToViewModel(this Location location) to AllReady.Areas.Admin.Models.LocationExtensions and deleting the newly added LocationExtensions class from Lcoation.cs
